ZUT25 check for available starting places for each track

ZUT25 check for available starting places for each track

Usage

check_starting_places(url, send_pushover_notification = FALSE, dbg = TRUE)

Arguments

url

one element of zut25_urls

send_pushover_notification

should pushover notifiation be sent (default: FALSE)

dbg

print debug messages (default: TRUE)

Value

either 0 (if no starting place is available) or number of available places as text. in additon an email will be sent to recipient email in case a starting place is available

Examples

sapply(zut25_urls(), check_starting_places)
#> 
#> 
#> 
#> 
#> 
#> 
#> zut25_ultratrail    zut25_ehrwald   zut25_leutasch zut25_mittenwald 
#>                0                0                0                0 
#>   zut25_garmisch    zut25_grainau 
#>                0                0